requirements.txt
# Sakshyam Audio Intelligence Backend
# Install: pip install -r requirements.txt

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Audio transcription (runs 100% locally, no API key)
openai-whisper==20231117

# Utilities
aiofiles==23.2.1
